requirements.txt
boto3==1.18.3
botocore==1.21.3
greenlet==1.1.0
jmespath==0.10.0
numpy==1.21.1
pandas==1.3.0
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.21
urllib3==1.26.6
